WIP
I think I had too many objectives with emit_function(), so I've reduced its
scope to figuring out which "substances" (sub-instance) are required, and
constructing an execution plan of sorts, consisting of which subinstances to
execute and when (which is the "hard part of the execution plan" anyway)

function emission is a bit early at this point, because there are a number of
analyses that can change the number of functions, what they do exactly, and
their "signatures" (arguments and return values). my previous approach only
worried about anti-rematerialization, and didn't even solve the problem really.
but things like constant folding (hardcoding arguments) and inlining also
changes functions... and constant folding seems related to The Primary
Challenge of reducing the complexity of "ram chip simulation" to something
manageable; it'll likely be something like, "if this input has that value, use
this function, otherwise that function", which is related to constant folding
in the sense that function arguments are hardcoded

I also think that I should be able to convert the product of
produce_substance() to something executable without too much trouble, even
though it's not optimized at all. getting the code to a state where I can start
executing stuff is important; test cases can verify correctness while I hack on
optimizations. it's a lot harder to do "blind"