freeing tmp data considered HARD
to free tmp data (drains, outcomes, queue,...) I have to relocate the sequence
array as well as outcome_request_bs32_p in new zvm_substance_key values...
which is hard because zvm_substance_key is duplicated, and so I have to binary
search keys while I'm mutating them.

ON SECOND THOUGHT why not simply consider the tmp data allocations lost? :) in
the grand scheme, buf is meant to be temporary anyway; a scratch buffer that is
freed when compilation is done.