calculate lut size...
it's how many bits that are required to represent the entire substance as a
look-up-table... for low numbers (e.g. 32 or less) it makes a lot of sense to
simply emit the LUT instead of the code inside... modules like NOT,AND,OR,etc
are definitely within this range (sz=4 for binary ops)... even memory_bit fits
within 16 bits... a 4-to-16 decoder fits inside 256 bits (16 outputs, and 16
possible inputs, 16*16=256) which is 8 u32 words... so while it might be better
NOT to inline a 4-to-16 decoder, it's still interesting to reduce the function
to its LUT, because naively a 4-to-16 decoder requires at least 4*16
operations? more like 6*16 if I fail to optimize the redundant NOT(input)
operations